+++ app/components/role-type-options/template.hbs
@@ -1,0 +1,15 @@
+<div class="form-group">
+  <label>Type</label>
+  <div class="role-type-select role-type-select--platform role-type-select--selected">
+    <h3 class="role-type-select__name">Platform</h3>
+    <p class="role-type-select__info">
+      For granting permission to Apps, Databases, and Environments
+    </p>
+  </div>
+  <div class="role-type-select role-type-select--compliance">
+    <h3 class="role-type-select__name">Compliance</h3>
+    <p class="role-type-select__info">
+      For granting permission to Compliance Products
+    </p>
+  </div>
+</div>
