+++ app/organization/roles/index/route.js
@@ -13,8 +13,6 @@ export default Ember.Route.extend({
   afterModel(model){
     return Ember.RSVP.hash({
       users: Ember.RSVP.all(model.roles.map(r => r.get('users'))),
-      // Preloading stack permissions for speedier rendering
-      // Permissions are embedded in stacks#index
       permissions: Ember.RSVP.all(model.stacks.map(s => s.get('permissions')))
     });
   },
@@ -23,6 +21,5 @@ export default Ember.Route.extend({
     controller.set('model', model.roles);
     controller.set('stacks', model.stacks);
     controller.set('organization', this.modelFor('organization'));
-    controller.set('ownerRole', this.modelFor('organization').ownerRole);
   }
 });
