+++ app/organization/roles/index/template.hbs
@@ -1,12 +1,36 @@
 {{partial 'organization/roles-header'}}
 
 <div class="layout-container">
+
+  <div class="row">
+    <div class="col-xs-12">
+      {{#aptible-ability user=session.currentUser scope='manage' permittable=organization as |hasAbility|}}
+        {{#if hasAbility}}
+          {{#link-to 'role' ownerRole title=(concat-params "Edit " ownerRole.name " Permissions") classNames="panel-link"}}
+            {{!partial 'organization/role' role=ownerRole}}
+          {{/link-to}}
+        {{else}}
+          {{!partial 'organization/role' role=ownerRole}}
+        {{/if}}
+      {{/aptible-ability}}
+    </div>
+  </div>
+
+  <ul class='nav nav-pills resource-navigation'>
+    {{#activating-item currentWhen="organization.roles" tagName="li"}}
+      {{link-to "Platform Roles" role "organization.roles"}}
+    {{/activating-item}}
+    {{#activating-item currentWhen="organization.roles.compliance" tagName="li"}}
+      {{link-to "Compliance Roles" role "organization.roles.compliance"}}
+    {{/activating-item}}
+  </ul>
+
   <div class="row">
     {{#each sortedRoles as |role|}}
       <div class="col-xs-12">
         {{#aptible-ability user=session.currentUser scope='manage' permittable=organization as |hasAbility|}}
           {{#if hasAbility}}
-            {{#link-to 'organization.roles.show' role title=(concat-params "Edit " role.name " Permissions") classNames="panel-link"}}
+            {{#link-to 'role' role title=(concat-params "Edit " role.name " Permissions") classNames="panel-link"}}
               {{partial 'organization/role'}}
             {{/link-to}}
           {{else}}
