+++ app/organization/roles/new/route.js
@@ -1,50 +1,22 @@
 import Ember from 'ember';
-import Changeset from "diesel/utils/changeset";
 
 export default Ember.Route.extend({
-  model(){
+  model() {
     let organization = this.modelFor('organization');
     return this.store.createRecord('role', {
       organization
     });
   },
 
-  afterModel(){
+  afterModel() {
     let organization = this.modelFor('organization');
-    return this.store.findStacksFor(organization).then((stacks) => {
-      this._stacks = stacks;
-    });
+    return Ember.RSVP.resolve(organization.get('billingDetail'));
   },
 
-  setupController(controller, model){
-    this._super(controller, model);
-    controller.set('stacks', this._stacks);
-    controller.set('organization', this.modelFor('organization'));
-
-    const changeset = Changeset.create({
-      key(keyData) {
-        return `${keyData.stack.get('handle')}-${keyData.scope}`;
-      },
-      initialValue(keyData) {
-        const permissions = keyData.stack.get('permissions.content');
-        const role = keyData.role;
-        const scope = keyData.scope;
-
-        const value = {};
-
-        value.permission = permissions.find((perm) => {
-          let roleLink = role.get('data.links.self');
-          let permRoleLink = perm.get('data.links.role');
-          return roleLink === permRoleLink &&
-            perm.get('scope') === scope;
-        });
-        value.isEnabled = !!value.permission;
-
-        return value;
-      }
-    });
-    controller.set('changeset', changeset);
-    //controller.observeChangeset();
+  setupController(controller, model) {
+    let organization = this.modelFor('organization');
+    controller.set('organization', organization);
+    controller.set('billingDetail', organization.get('billingDetail'));
   },
 
   actions: {
@@ -58,29 +30,13 @@ export default Ember.Route.extend({
 
     save(){
       let role = this.currentModel;
-      let changeset = this.controller.get('changeset');
 
       role.save().then(() => {
-        let newPermissions = [];
-        changeset.forEachValue((keyData, initialValue, value) => {
-          if (value.isEnabled) {
-            let permission = this.store.createRecord('permission', {
-              stack: keyData.stack,
-              role:  keyData.role.get('data.links.self'),
-              scope: keyData.scope
-            });
-            newPermissions.push( permission.save() );
-          }
-        });
-
-        return Ember.RSVP.all(newPermissions);
-      }).then(() => {
         let message = `${role.get('name')} created`;
 
-        this.transitionTo('organization.roles.index');
+        this.transitionTo('role.environments', role);
         Ember.get(this, 'flashMessages').success(message);
       });
     }
   }
 });
-
