+++ app/role/index/route.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  redirect: function(){
+    this.transitionTo('role.members');
+  }
+});
