+++ app/role/route.js
@@ -1,0 +1,11 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  renderTemplate() {
+    this._super.apply(this, arguments);
+    this.render('sidebars/settings', {
+      into: 'dashboard',
+      outlet: 'sidebar'
+    });
+  }
+});
