+++ app/role/settings/controller.js
@@ -1,4 +1,4 @@
-import Ember from 'ember';
+import Ember from "ember";
 
 export default Ember.Controller.extend({
   isUnchanged: Ember.computed.not('model.isDirty')
