+++ app/role/template.hbs
@@ -1,0 +1,29 @@
+<main class="layout-content">
+  <header class="resource-header">
+    <div class="layout-container">
+      <div class="resource-title">
+        <h1 class="resource-breadcrumb">
+          <span class="current">{{model.name}}</span>
+        </h1>
+      </div>
+    </div>
+  </header>
+
+  <div class="layout-container">
+
+    <ul class='nav nav-pills resource-navigation'>
+      {{#activating-item currentWhen='role.members' tagName='li'}}
+        {{link-to 'Members' 'role.members' model}}
+      {{/activating-item}}
+      {{#activating-item currentWhen='role.environments' tagName='li'}}
+        {{link-to 'Environments' 'role.environments' model}}
+      {{/activating-item}}
+      {{#activating-item currentWhen='role.settings' tagName='li'}}
+        {{link-to 'Settings' 'role.settings' model}}
+      {{/activating-item}}
+    </ul>
+
+    {{outlet}}
+
+  </div>
+</main>
