+++ app/styles/app.scss
@@ -26,6 +26,7 @@
 @import "components/payment-form";
 @import "components/resource-icon";
 @import "components/role-members";
+@import "components/role-type-options";
 @import "components/save-progress";
 @import "components/service-badge";
 @import "components/service-scaler";
