+++ app/styles/components/_labels.scss
@@ -16,3 +16,6 @@
 .label-success {
   background-color: #7ED321;
 }
+.label-info {
+  background-color: #9B9B9B;
+}
