+++ app/styles/components/_panel.scss
@@ -27,12 +27,13 @@
   }
 
   .panel-body--aptable {
+    border-left: 1px solid #E0E0E0;
     padding: 0;
     text-align: left;
     .aptable {
       border-width: 0;
+      border-bottom-width: 1px;
       box-shadow: none;
-      border-left-width: 1px;
       border-radius: 0;
       // thead td {
       //   font-size: 12px;
@@ -86,5 +87,4 @@
 .user-names__value, .user-names__remainder {
   font-size: 12px;
   color: #9B9B9B;
-  white-space: nowrap;
 }
