+++ app/styles/components/slider.scss
@@ -151,6 +151,7 @@
     margin-top: 12px;
 
     .label {
+      background-color: transparent;
       color: $color-light;
       text-transform: uppercase;
       font-size: 12px;
