+++ app/templates/sidebars/settings.hbs
@@ -11,9 +11,9 @@
         {{link-to "Pending Invitations" "organization.members.pending-invitations" organization bubbles=false}}
       {{/activating-item}}
     {{/if}}
-    {{#link-to 'organization.roles' organization tagName="li"}}
+    {{#activating-item currentWhen="organization.role" tagName="li"}}
       {{link-to "Roles" "organization.roles" organization bubbles=false}}
-    {{/link-to}}
+    {{/activating-item}}
     {{#link-to 'organization.environments' organization tagName="li"}}
       {{link-to "Environments" "organization.environments" organization bubbles=false}}
     {{/link-to}}
